Use local time zone so :hours are in correct zone
Cron jobs are usually in the local time zone. Chronicle by default produces times that are in UTC, so {:hours [0]} would be 7am or 8am here on the left coast of the USA, for example, not midnight as intended. 

Make moments use the local time zone for jobs to be run a certain hours.